backend/requirements.txt
setuptools
Django
numpy
mysqlclient
djangorestframework
PyJWT
bcrypt
datetime
django-cors-headers